requirements.txt
streamlit
transformers
torch
pytesseract
Pillow
